database_init/requirements.txt
mysql-connector-python
python-dotenv
httpx
pytest
